Extract MobileMenu
