+++ site/src/modules/dashboard/Navbar/NavbarView.tsx
@@ -2,36 +2,13 @@ import type * as TypesGen from "api/typesGenerated";
 import { ExternalImage } from "components/ExternalImage/ExternalImage";
 import { CoderIcon } from "components/Icons/CoderIcon";
 import type { ProxyContextValue } from "contexts/ProxyContext";
-import { useState, type FC } from "react";
-import { Link, NavLink, useLocation } from "react-router-dom";
+import type { FC } from "react";
+import { NavLink, useLocation } from "react-router-dom";
 import { DeploymentDropdown } from "./DeploymentDropdown";
 import { ProxyMenu } from "./ProxyMenu";
 import { UserDropdown } from "./UserDropdown/UserDropdown";
 import { cn } from "utils/cn";
-import { Button } from "components/Button/Button";
-import {
-	ChevronRightIcon,
-	CircleHelpIcon,
-	MenuIcon,
-	XIcon,
-} from "lucide-react";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from "components/DropdownMenu/DropdownMenu";
-import { Avatar } from "components/Avatar/Avatar";
-import { Latency } from "components/Latency/Latency";
-import {
-	Collapsible,
-	CollapsibleContent,
-	CollapsibleTrigger,
-} from "components/Collapsible/Collapsible";
-import { sortProxiesByLatency } from "./proxyUtils";
-import { displayError } from "components/GlobalSnackbar/utils";
-import { FeatureStageBadge } from "components/FeatureStageBadge/FeatureStageBadge";
+import { MobileMenu } from "./MobileMenu";
 
 export interface NavbarViewProps {
 	logo_url?: string;
@@ -48,16 +25,10 @@ export interface NavbarViewProps {
 	proxyContextValue?: ProxyContextValue;
 }
 
-const linkClassNames = {
+const linkStyles = {
 	default:
 		"text-sm font-medium text-content-secondary no-underline block h-full px-2 flex items-center hover:text-content-primary transition-colors",
 	active: "text-content-primary",
-};
-
-const mobileDropdownItemClassNames = {
-	default: "px-9 h-10 no-underline",
-	sub: "pl-12",
-	open: "text-content-primary",
 };
 
 export const NavbarView: FC<NavbarViewProps> = ({
@@ -100,7 +71,7 @@ export const NavbarView: FC<NavbarViewProps> = ({
 				/>
 
 				<a
-					className={linkClassNames.default}
+					className={linkStyles.default}
 					href={docsHref}
 					target="_blank"
 					rel="noreferrer"
@@ -129,323 +100,6 @@ export const NavbarView: FC<NavbarViewProps> = ({
 	);
 };
 
-type MobileMenuProps = {
-	proxyContextValue?: ProxyContextValue;
-	user?: TypesGen.User;
-	supportLinks?: readonly TypesGen.LinkConfig[];
-	docsHref: string;
-	onSignOut: () => void;
-};
-
-const MobileMenu: FC<MobileMenuProps> = ({
-	proxyContextValue,
-	user,
-	supportLinks,
-	docsHref,
-	onSignOut,
-}) => {
-	const [open, setOpen] = useState(false);
-
-	return (
-		<DropdownMenu open={open} onOpenChange={setOpen}>
-			{open && (
-				<div className="fixed inset-0 top-[72px] backdrop-blur-sm z-10 bg-content-primary/50" />
-			)}
-			<DropdownMenuTrigger asChild>
-				<Button
-					aria-label={open ? "Close menu" : "Open menu"}
-					size="icon"
-					variant="ghost"
-					className="ml-auto md:hidden [&_svg]:size-6"
-				>
-					{open ? (
-						<XIcon className="size-icon-lg" />
-					) : (
-						<MenuIcon className="size-icon-lg" />
-					)}
-				</Button>
-			</DropdownMenuTrigger>
-			<DropdownMenuContent
-				className="w-screen border-0 border-b border-solid p-0 py-2"
-				sideOffset={17}
-			>
-				<ProxySettingsSub proxyContextValue={proxyContextValue} />
-				<DropdownMenuSeparator />
-				<AdminSettingsSub />
-				<DropdownMenuSeparator />
-				<DropdownMenuItem
-					asChild
-					className={mobileDropdownItemClassNames.default}
-				>
-					<a href={docsHref} target="_blank" rel="noreferrer norefereer">
-						Docs
-					</a>
-				</DropdownMenuItem>
-				<DropdownMenuSeparator />
-				<UserSettingsSub
-					user={user}
-					supportLinks={supportLinks}
-					onSignOut={onSignOut}
-				/>
-			</DropdownMenuContent>
-		</DropdownMenu>
-	);
-};
-
-type ProxySettingsSubProps = {
-	proxyContextValue?: ProxyContextValue;
-};
-
-const ProxySettingsSub: FC<ProxySettingsSubProps> = ({ proxyContextValue }) => {
-	const selectedProxy = proxyContextValue?.proxy.proxy;
-	const latency = selectedProxy
-		? proxyContextValue?.proxyLatencies[selectedProxy?.id]
-		: undefined;
-	const [open, setOpen] = useState(false);
-
-	if (!selectedProxy) {
-		return null;
-	}
-
-	return (
-		<Collapsible open={open} onOpenChange={setOpen}>
-			<CollapsibleTrigger asChild>
-				<DropdownMenuItem
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						open ? mobileDropdownItemClassNames.open : "",
-					)}
-					onClick={(e) => {
-						e.preventDefault();
-						setOpen((prev) => !prev);
-					}}
-				>
-					Workspace proxy settings:
-					<span className="leading-[0px] flex items-center gap-1">
-						<img
-							className="w-4 h-4"
-							src={selectedProxy.icon_url}
-							alt={selectedProxy.name}
-						/>
-						{latency && <Latency latency={latency.latencyMS} />}
-					</span>
-					<ChevronRightIcon
-						className={cn("ml-auto", open ? "rotate-90" : "")}
-					/>
-				</DropdownMenuItem>
-			</CollapsibleTrigger>
-			<CollapsibleContent>
-				{proxyContextValue.proxies &&
-					sortProxiesByLatency(
-						proxyContextValue.proxies,
-						proxyContextValue.proxyLatencies,
-					).map((p) => {
-						const latency = proxyContextValue.proxyLatencies[p.id];
-						return (
-							<DropdownMenuItem
-								className={cn(
-									mobileDropdownItemClassNames.default,
-									mobileDropdownItemClassNames.sub,
-								)}
-								key={p.id}
-								onClick={(e) => {
-									e.preventDefault();
-
-									if (!p.healthy) {
-										displayError("Please select a healthy workspace proxy.");
-										return;
-									}
-
-									proxyContextValue.setProxy(p);
-									setOpen(false);
-								}}
-							>
-								<img className="w-4 h-4" src={p.icon_url} alt={p.name} />
-								{p.display_name || p.name}
-								{latency ? (
-									<Latency latency={latency.latencyMS} />
-								) : (
-									<CircleHelpIcon className="ml-auto" />
-								)}
-							</DropdownMenuItem>
-						);
-					})}
-				<DropdownMenuSeparator />
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/deployment/workspace-proxies">Proxy settings</Link>
-				</DropdownMenuItem>
-				<DropdownMenuItem
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-					onClick={() => {
-						proxyContextValue.refetchProxyLatencies();
-					}}
-				>
-					Refresh latencies
-				</DropdownMenuItem>
-			</CollapsibleContent>
-		</Collapsible>
-	);
-};
-
-const AdminSettingsSub: FC = () => {
-	const [open, setOpen] = useState(false);
-
-	return (
-		<Collapsible open={open} onOpenChange={setOpen}>
-			<CollapsibleTrigger asChild>
-				<DropdownMenuItem
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						open ? mobileDropdownItemClassNames.open : "",
-					)}
-					onClick={(e) => {
-						e.preventDefault();
-						setOpen((prev) => !prev);
-					}}
-				>
-					Admin settings
-					<ChevronRightIcon
-						className={cn("ml-auto", open ? "rotate-90" : "")}
-					/>
-				</DropdownMenuItem>
-			</CollapsibleTrigger>
-			<CollapsibleContent>
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/deployment/general">Deployment</Link>
-				</DropdownMenuItem>
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/organizations">
-						Organizations
-						<FeatureStageBadge
-							contentType="beta"
-							size="sm"
-							showTooltip={false}
-						/>
-					</Link>
-				</DropdownMenuItem>
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/audit">Audit logs</Link>
-				</DropdownMenuItem>
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/health">Healthcheck</Link>
-				</DropdownMenuItem>
-			</CollapsibleContent>
-		</Collapsible>
-	);
-};
-
-type UserSettingsSubProps = {
-	user?: TypesGen.User;
-	supportLinks?: readonly TypesGen.LinkConfig[];
-	onSignOut: () => void;
-};
-
-const UserSettingsSub: FC<UserSettingsSubProps> = ({
-	user,
-	supportLinks,
-	onSignOut,
-}) => {
-	const [open, setOpen] = useState(false);
-
-	return (
-		<Collapsible open={open} onOpenChange={setOpen}>
-			<CollapsibleTrigger asChild>
-				<DropdownMenuItem
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						open ? mobileDropdownItemClassNames.open : "",
-					)}
-					onClick={(e) => {
-						e.preventDefault();
-						setOpen((prev) => !prev);
-					}}
-				>
-					<Avatar
-						src={user?.avatar_url}
-						fallback={user?.name || user?.username}
-					/>
-					User settings
-					<ChevronRightIcon
-						className={cn("ml-auto", open ? "rotate-90" : "")}
-					/>
-				</DropdownMenuItem>
-			</CollapsibleTrigger>
-			<CollapsibleContent>
-				<DropdownMenuItem
-					asChild
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-				>
-					<Link to="/settings/account">Account</Link>
-				</DropdownMenuItem>
-				<DropdownMenuItem
-					className={cn(
-						mobileDropdownItemClassNames.default,
-						mobileDropdownItemClassNames.sub,
-					)}
-					onClick={onSignOut}
-				>
-					Sign out
-				</DropdownMenuItem>
-				{supportLinks && (
-					<>
-						<DropdownMenuSeparator />
-						{supportLinks?.map((l) => (
-							<DropdownMenuItem
-								key={l.name}
-								asChild
-								className={cn(
-									mobileDropdownItemClassNames.default,
-									mobileDropdownItemClassNames.sub,
-								)}
-							>
-								<a href={l.target} target="_blank" rel="noreferrer">
-									{l.name}
-								</a>
-							</DropdownMenuItem>
-						))}
-					</>
-				)}
-			</CollapsibleContent>
-		</Collapsible>
-	);
-};
-
 interface NavItemsProps {
 	className?: string;
 }
@@ -460,10 +114,7 @@ const NavItems: FC<NavItemsProps> = ({ className }) => {
 					if (location.pathname.startsWith("/@")) {
 						isActive = true;
 					}
-					return cn(
-						linkClassNames.default,
-						isActive ? linkClassNames.active : "",
-					);
+					return cn(linkStyles.default, isActive ? linkStyles.active : "");
 				}}
 				to="/workspaces"
 			>
@@ -471,10 +122,7 @@ const NavItems: FC<NavItemsProps> = ({ className }) => {
 			</NavLink>
 			<NavLink
 				className={({ isActive }) => {
-					return cn(
-						linkClassNames.default,
-						isActive ? linkClassNames.active : "",
-					);
+					return cn(linkStyles.default, isActive ? linkStyles.active : "");
 				}}
 				to="/templates"
 			>
